+++ Changelog.md
@@ -1,14 +1,19 @@
 # Change Log
 
-## 0.1.1
+## 0.2.0
 
 Breaking Changes
+
+*  `PrimitiveContent`’s methods take `self` instead of `&self`. This
+   avoids the lifetime argument in `Primitive`, its encoder. [(#6)]
 
 New
 
 Bug Fixes
 
 Dependencies
+
+[(#6)]: https://github.com/NLnetLabs/bcder/pull/6
 
 
 ## 0.1.0